Original Roberts Rules Controller
This is the first crude attempt which mostly uses client-side
javascript. I later added a bit of query. The most impressive part is
the Q and A Controller, which includes countdown clocks and the moving
of a resolution (principal motion) as well as a few other Roberts Rules
bells and whistles.